+++ tools/setup_environment.ps1
@@ -9,6 +9,8 @@ try {
     if (-not (Test-Path -LiteralPath $forgegatePython)) {
         py -3.12 -m venv .venv
     }
+    & $forgegatePython -m pip install --disable-pip-version-check --upgrade `
+        -c $forgegateConstraints pip
     & $forgegatePython -m pip install --disable-pip-version-check `
         -c $forgegateConstraints -e ".[dev]"
     & $forgegatePython tools\verify.py